+++ os_doc_tools/doctest.py
@@ -50,6 +50,7 @@ from lxml import etree
 from oslo_config import cfg
 
 import os_doc_tools
+from os_doc_tools import index
 from os_doc_tools import jsoncheck
 from os_doc_tools.openstack.common import log
 
@@ -191,7 +192,7 @@ def verify_attribute_profiling(doc, attribute, known_values):
                         "profiling of %s on line %d." %
                         (p_tag[len_ns:], attribute, p_att_list,
                          attribute, c_tag[len_ns:], c_line))
-    if len(msg) > 0:
+    if msg:
         raise ValueError("\n     ".join(msg))
 
 
@@ -223,7 +224,7 @@ def verify_unicode_niceness(docfile):
         lc += 1
         any(c in line for c in invalid_characters)
 
-    if len(affected_lines):
+    if affected_lines:
         raise ValueError("unwanted unicode charaters (one of %s) "
                          "found in line(s): %" % (" ".join(invalid_characters),
                                                   ", ".join(affected_lines)))
@@ -321,7 +322,7 @@ def verify_valid_links(doc):
             msg.append("URL %s invalid at line %d, error %s" % (
                 url, e_line, e))
 
-    if len(msg) > 0:
+    if msg:
         raise ValueError("\n    ".join(msg))
 
 
@@ -972,36 +973,6 @@ def build_book(book, publish_path, log_path):
                 ["mvn", "generate-sources", comments, release, "-B"],
                 stderr=subprocess.STDOUT
             )
-        # Repository: identity-api
-        elif (cfg.CONF.repo_name == "identity-api"
-              and book.endswith("v3")):
-            output = subprocess.check_output(
-                ["bash", os.path.join(SCRIPTS_DIR, "markdown-docbook.sh"),
-                 "identity-api-v3"],
-                stderr=subprocess.STDOUT
-            )
-            out_file.write(output)
-            # File gets generated at wrong directory, we need to move it
-            # around
-            if os.path.isfile('identity-api-v3.xml'):
-                os.remove('identity-api-v3.xml')
-            shutil.move("src/markdown/identity-api-v3.xml", ".")
-            output = subprocess.check_output(
-                ["mvn", "generate-sources", comments, release, "-B"],
-                stderr=subprocess.STDOUT
-            )
-        # Repository: image-api
-        elif base_book == 'image-api-v2':
-            output = subprocess.check_output(
-                ["bash", os.path.join(SCRIPTS_DIR, "markdown-docbook.sh"),
-                 "image-api-v2.0"],
-                stderr=subprocess.STDOUT
-            )
-            out_file.write(output)
-            output = subprocess.check_output(
-                ["mvn", "generate-sources", comments, release, "-B"],
-                stderr=subprocess.STDOUT
-            )
         else:
             output = subprocess.check_output(
                 ["mvn", "generate-sources", comments, release, "-B"],
@@ -1093,7 +1064,7 @@ def generate_affected_books(rootdir, book_bk, ignore_dirs, abs_ignore_dirs,
     # 3. Iterate over files that have includes on modified files
     # and build a closure - the set of all files (affected_files)
     # that have a path to a modified file via includes.
-    while len(new_files) > 0:
+    while new_files:
         new_files_to_check = new_files
         new_files = []
         for f in new_files_to_check:
@@ -1241,67 +1212,6 @@ def find_affected_books(rootdir, book_exceptions, file_exceptions,
     return books
 
 
-def generate_index_file():
-    """Generate index.html file in publish_path."""
-
-    publish_path = get_publish_path()
-    if not os.path.isdir(publish_path):
-        os.mkdir(publish_path)
-
-    index_file = open(os.path.join(get_publish_path(), 'index.html'), 'w')
-
-    index_file.write(
-        '<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN"\n'
-        '"http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">\n'
-        '<html lang="en" xmlns="http://www.w3.org/1999/xhtml" xml:lang="en">\n'
-        '<body>\n'
-        '<h1>Results of checkbuild</h1>\n')
-
-    links = {}
-    for root, dirs, files in os.walk(publish_path):
-
-        dirs[:] = [d for d in dirs if d not in ['common', 'webapp', 'content',
-                                                'www']]
-
-        # Ignore top-level index.html files
-        if root == publish_path:
-            continue
-
-        if os.path.isfile(os.path.join(root, 'content/index.html')):
-            path = os.path.relpath(root, publish_path)
-            links[path] = ('<a href="%s/content/index.html">%s</a>\n' %
-                           (path, path))
-        elif os.path.isfile(os.path.join(root, 'index.html')):
-            path = os.path.relpath(root, publish_path)
-            links[path] = ('<a href="%s/index.html">%s</a>\n' %
-                           (path, path))
-
-        if os.path.isfile(os.path.join(root, 'api-ref.html')):
-            path = os.path.relpath(root, publish_path)
-            links[path] = ('<a href="%s/api-ref.html">%s</a>\n' %
-                           (path, path))
-
-        # List PDF files for api-site that have from "bk-api-ref*.pdf"
-        # as well since they have no corresponding html file.
-        for f in files:
-            if f.startswith('bk-api-ref') and f.endswith('.pdf'):
-                path = os.path.relpath(root, publish_path)
-                links[f] = ('<a href="%s/%s">%s</a>\n' %
-                            (path, f, f))
-
-    for entry in sorted(links):
-        index_file.write(links[entry])
-        index_file.write('<br/>\n')
-
-    if os.path.isfile(os.path.join(get_publish_path(), 'www-index.html')):
-        index_file.write('<br/>\n')
-        index_file.write('<a href="www-index.html">list of generated '
-                         'WWW pages</a>\n')
-    index_file.write('</body>\n'
-                     '</html>\n')
-    index_file.close()
-
-
 def build_affected_books(rootdir, book_exceptions, file_exceptions,
                          force=False, ignore_dirs=None,
                          ignore_books=None):
@@ -1379,7 +1289,7 @@ def build_affected_books(rootdir, book_exceptions, file_exceptions,
             any_failures = True
 
     if cfg.CONF.create_index:
-        generate_index_file()
+        index.generate_index_file(get_publish_path())
 
     if any_failures:
         for book, result, output, returncode in RESULTS_OF_BUILDS:
@@ -1633,7 +1543,7 @@ def doctest():
 
     if check_no_building():
         if CONF.create_index:
-            generate_index_file()
+            index.generate_index_file(get_publish_path())
         return
 
     if CONF.check_syntax or CONF.check_niceness or CONF.check_links: