+++ sitemap/generator/spiders/sitemap.py
@@ -10,7 +10,6 @@
 # License for the specific language governing permissions and limitations
 # under the License.
 
-import posixpath
 import time
 import urlparse
 
@@ -21,6 +20,19 @@ from scrapy import spiders
 
 class SitemapSpider(spiders.CrawlSpider):
     name = 'sitemap'
+    old_releases = tuple(["/%s" % old_release for old_release in [
+        'austin',
+        'bexar',
+        'cactus',
+        'diablo',
+        'essex',
+        'folsom',
+        'grizzly',
+        'havana',
+        'icehouse',
+        'juno',
+        'kilo'
+    ]])
 
     rules = [
         spiders.Rule(
@@ -42,37 +54,34 @@ class SitemapSpider(spiders.CrawlSpider):
         )
     ]
 
-    def __init__(self, domain='docs.openstack.org', *args, **kwargs):
+    def __init__(self, domain='docs.openstack.org', urls='', *args, **kwargs):
         super(SitemapSpider, self).__init__(*args, **kwargs)
         self.domain = domain
         self.allowed_domains = [domain]
-        self.start_urls = [
-            'http://%s/index.html' % domain,
-        ]
+        self.start_urls = ['http://%s' % domain]
+        for url in urls.split(','):
+            if not url:
+                continue
+            self.start_urls.append(url)
 
     def parse_item(self, response):
         item = items.SitemapItem()
-        item['priority'] = '0.5'
-        item['changefreq'] = 'daily'
         item['loc'] = response.url
 
         path = urlparse.urlsplit(response.url).path
-        filename = posixpath.basename(path)
-
-        if filename == 'index.html' or filename == '':
+        if path.startswith(self.old_releases):
+            # weekly changefrequency and lower priority for old files
+            item['priority'] = '0.5'
+            item['changefreq'] = 'weekly'
+        else:
+            # daily changefrequency and highest priority for current files
             item['priority'] = '1.0'
+            item['changefreq'] = 'daily'
 
-        weekly = [
-            'juno',
-            'icehouse',
-            'havana'
-        ]
-
-        for entry in weekly:
-            if path.startswith("/%s" % entry):
-                item['changefreq'] = 'weekly'
-
-        lastmod = time.strptime(response.headers['Last-Modified'],
-                                "%a, %d %b %Y %H:%M:%S %Z")
+        if 'Last-Modified' in response.headers:
+            timestamp = response.headers['Last-Modified']
+        else:
+            timestamp = response.headers['Date']
+        lastmod = time.strptime(timestamp, "%a, %d %b %Y %H:%M:%S %Z")
         item['lastmod'] = time.strftime("%Y-%m-%dT%H:%M:%S%z", lastmod)
         return item